Bump Dockerfiles + rename tests
