+++ airbyte-integrations/connectors/source-mysql/src/main/java/io/airbyte/integrations/source/mysql/initialsync/MySqlFeatureFlags.java
@@ -9,15 +9,10 @@ import com.fasterxml.jackson.databind.JsonNode;
 // Feature flags to gate new primary key load features.
 public class MySqlFeatureFlags {
 
-  public static final String CDC_VIA_PK = "cdc_via_pk";
   private final JsonNode sourceConfig;
 
   public MySqlFeatureFlags(final JsonNode sourceConfig) {
     this.sourceConfig = sourceConfig;
-  }
-
-  public boolean isCdcInitialSyncViaPkEnabled() {
-    return getFlagValue(CDC_VIA_PK);
   }
 
   private boolean getFlagValue(final String flag) {
